Return removed=true for non-resident channel removal revisions (#2821)
* Return removed=true for non-resident channel removal revisions

When an revision is requested, but that revision's body is no longer available (it's not resident in the rev cache, and the temporary revision backup has expired), check whether that revision is associated with a channel removal based on the document's channel history.  If it's a removal, return the standard removal notification body (_removed=true) instead of 404 missing.

* Add additional tests, code cleanup

Additional tests and code cleanup based on PR review.

* Add new tests to integration suite

# Conflicts:
#	test_integration.sh